+++ CHANGELOG.md
@@ -32,6 +32,10 @@ git logs & PR history.
 - support amazon as a platform
 - add codenames for MacOS 10.13 and Ubuntu Artful
 
+### Added
+
+- fine-grained, structured timing for tests
+
 # [3.26.0](https://github.com/puppetlabs/beaker/compare/3.25.0...3.26.0) - 2017-10-05
 
 ### Added